Additional.. Coursework-Uploads: Phillips-Silver Reading Guides and Drafts
This commit adds "three more" documents to the Coursework-Uploads folder, focusing on analyses and drafts which are related to Phillips-Silver et al.'s Born to Dance but Beat Deaf: A New Form of Congenital Amusia. The files are inclusive of the following: 

Gladish_Reading_Guide_Phillips_Silver_2011_Born_to_Dance.docx: A detailed reading guide analyzing the most concise findings of the study.
Gladish_RG4_Draft_Phillips_Silver_2011_Born_to_Dance.docx: The fourth draft of an essay exploring the relevance and pertinence of the study.
Gladish_RG4_Draft_Comments_Phillips_Silver_2011_Born_to_Dance.docx: A version with comments and significant feedback from Katherine. 
These files are a part to and of the academic exploration of congenital amusia and its broader implications and re-purposing in and via neuropsychology.